magento/adobe-stock-integration#1444: Save metadata to file on editing - updated PR with revisions
